Fixing web platform tests for PeerConnection.setRemoteDescription().
Tests for PeerConnection.setRemoteDescription() did not call
setLocalDescription() on the caller object after creating the offer.
When the local description is not set, the PeerConnection can generate new MIDs
for the existing media sections. This caused tests to fail because the offer
was set on the callee causing the MIDs (and therefore media sections) not to
match.

Bug: None.
Change-Id: I67e98e5abf76a97ab57fc0427672313b44f9c258